requirements.txt
pandas==2.0.3
python-dotenv==1.0.0
setuptools==65.5.0
SQLAlchemy==2.0.19
